+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 
 # Master (Unreleased)
 
+### Breaking Changes:
+
+- Use datetime2 for MSSQL datetime + timestamp types. This change is incompatible with MSSQL older than 2008 #2757
 - Introduced support for specifying multiple directories for the Migrator #2735
 
 # 0.15.2 - 19 Jul, 2018